% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxcox.t}
\alias{boxcox.t}
\title{Box-Cox Transformation (boxcox.t)}
\usage{
boxcox.t(your.object)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
Returns optimal lambda value and transformed column vector
}
\description{
Box-Cox Power Transformations from Non-Normal to Normality
}
\details{
Box-Cox transformation is usually applied in order to achieve modeling assumptions.  As stated earlier techniques such as Simple Linear Regression, Multiple Linear Regression, Logistic Regression and other Classification Techniques like Discriminant Analysis, DT, NN, etc., also requires data to be normally distributed.(E.g.) In ENERGY model building, the usage of ENERGY in any house, district or state will be right-skewed.  Using Box-Cox when the same is transformed into ‘NORMAL’ the assumption is met, at the same time after transformation that particular units will lose its original UNIT OF MEASUREMENT.
}

