% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.iv}
\alias{create.iv}
\title{Creating Indicator Variables (create.iv)}
\usage{
create.iv(col.pos.x, iv.type)
}
\arguments{
\item{Numeric}{column position x in the dataframe and the type of indicator variable For REFERENCE CODING use ‘0’ ,ForEFFECT CODING use ‘-1’,For NO Y-INTERCEPT use ‘1’}
}
\value{
Indicator Variables Columns
}
\description{
Creates three types of indicator variables
}
\details{
Indicator Variables is a method that is widely used in modeling particularly to handle data that are NOMINAL in nature.  It converts NOMINAL data into a meaningful numerical variables knows as ‘Indicator Variables’ (IV).  The other names for Indicator Variables are also known as Dummy Variables (DV).At times, Indicator Variables (IV) were rarely applied even on NOMINAL data, to convert them into meaningful numerical variables.Indicator Variables (IVs) is widely applied in Regression Techniques, Machine Learning Techniques & Multivariate Techniques such as Cluster Analysis, etc.
}

