% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gs}
\alias{gs}
\title{Graphical Summary(gs)}
\usage{
gs(col.pos.x)
}
\arguments{
\item{Numeric}{(column position x in the dataframe)}
}
\value{
plots
}
\description{
Gives four types of graphical summary Histogram,Box plot,Time series plot and Auto correlation function plot
}
\details{
Histogram displays the values in Frequency Domain.  Histogram also displays information on Anderson-Darling Normality P-value, Mean (mu) & Standard Deviation (sigma).Time Series Plot, displays the values over Time Domain.  Time Plot displays values of Q1, Q2/Median & Q3.  These values are also displayed in the plot in the form of Horizontal Lines.Box-Whisker Plot, is a Five-Number summary.  Outliers and Skewness of the distribution can be assed using Box-Plot.Auto Correlation Function, shows the significant of each lags.
}

