% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mad.d}
\alias{mad.d}
\title{mean absolute Deviations (mad.d)}
\usage{
mad.d(col.poss)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
mean absolute deviation value
}
\description{
This functions gives the mean absolute deviation of a vector
}

