% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msd.d}
\alias{msd.d}
\title{mean square Deviations (msd.d)}
\usage{
msd.d(col.poss)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
mean square deviation value
}
\description{
This functions gives the mean square deviation of a vector
}

