% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my.deviate}
\alias{my.deviate}
\title{Deviations (my.deviate)}
\usage{
my.deviate(col.pos.x, Deviate = 0, MeanAbsDeviation = 1, MeanSqDev = 0)
}
\arguments{
\item{Four}{numeric parameters Parameter 1: It is a column of a Data Frame,Parameter 2: If the second parameter = 1, it then displays DEVIATIONS for each observation,Parameter 3: If the third parameter = 1, it then displays column required to compute MAD,Parameter 4: If the fourth parameter = 1, it then displays column required to compute MSD}
}
\value{
Displays the computed columns
}
\description{
This functions addresses Deviation, MAD & MSD.
}

