% Generated by roxygen2 (4.0.2): do not edit by hand
\name{poi.trans}
\alias{poi.trans}
\title{Transformation of Poisson Data (poi.trans)}
\usage{
poi.trans(your.object)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
Returns transformed column vector
}
\description{
Freeman Tukey Variance Stabilization transformation for POISSON DISTRIBUTION DATA
}
\details{
This technique widely used to Stabilize Variance during model building.if the COUNT data follows Poisson Distribution, this transformation is then applied.
}

