% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ske}
\alias{ske}
\title{Skewness (ske)}
\usage{
ske(column.pos)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
Gives the skewness value
}
\description{
Measure of Descriptive Statistics
}
\details{
Skewness addresses the characteristics of symmetry.
}

