% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stdZ}
\alias{stdZ}
\title{XtoZ (stdZ)}
\usage{
stdZ(pos.x, replace = 0)
}
\arguments{
\item{Numeric}{column position x in the dataframe}
}
\value{
Returns standardized column
}
\description{
Standardize
}
\details{
Transforms originalvalue into its respective "Z" value
}

